fix:290 pass, 35 fail

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/FilmController.java
@@ -3,12 +3,12 @@ package ru.yandex.practicum.filmorate.controllers;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import ru.yandex.practicum.filmorate.exceptions.AlreadyExistsException;
 import ru.yandex.practicum.filmorate.exceptions.FailedValidationException;
 import ru.yandex.practicum.filmorate.exceptions.InvalidInputException;
 import ru.yandex.practicum.filmorate.exceptions.NoSuchEntryException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.services.FilmService;
+import ru.yandex.practicum.filmorate.services.UserService;
 import ru.yandex.practicum.filmorate.utils.Messages;
 import ru.yandex.practicum.filmorate.utils.Validator;
 
@@ -20,10 +20,12 @@ import java.util.Optional;
 @RequestMapping("/films")
 public class FilmController {
     private final FilmService filmService;
+    private final UserService userService;
 
     @Autowired
-    public FilmController(FilmService filmService) {
+    public FilmController(FilmService filmService, UserService userService) {
         this.filmService = filmService;
+        this.userService = userService;
     }
 
     @GetMapping
@@ -33,7 +35,7 @@ public class FilmController {
 
     @GetMapping({"{id}"})
     public Optional<Film> getFilm(@PathVariable(value = "id") int filmID) {
-        if (filmService.filmIsPresent(filmID)) {
+        if (!filmService.filmIsPresent(filmID)) {
             throw new NoSuchEntryException(Messages.NO_SUCH_FILM);
         }
         return filmService.getFilm(filmID);
@@ -52,7 +54,7 @@ public class FilmController {
     public Film updateFilm(@RequestBody Film film) throws InvalidInputException {
         if (!Validator.isValidFilm(film)) {
             throw new FailedValidationException(Messages.FAILED_FILM_VALIDATION);
-        } else if (filmService.filmIsPresent(film.getId())) {
+        } else if (!filmService.filmIsPresent(film.getId())) {
             throw new NoSuchEntryException(Messages.NO_SUCH_FILM);
         }
         filmService.modifyFilm(film);
@@ -62,19 +64,28 @@ public class FilmController {
     @PutMapping("{id}/like/{userId}")
     public void likeFilm(@PathVariable(value = "id") int filmID,
                          @PathVariable(value = "userId") int userID) {
-        filmService.checkFilmAndUserPresence(filmID, userID);
-        filmService.addLike(filmID, userID);
+        if (filmService.filmIsPresent(filmID) & userService.userIsPresent(userID)) {
+            filmService.addLike(filmID, userID);
+        } else {
+            throw new NoSuchEntryException(Messages.TRY_ANOTHER_ID);
+        }
     }
 
     @DeleteMapping("{id}/like/{userId}")
     public void unlikeFilm(@PathVariable(value = "id") int filmID,
                            @PathVariable(value = "userId") int userID) {
-        filmService.checkFilmAndUserPresence(filmID, userID);
-        filmService.removeLike(filmID, userID);
+
+        if (filmService.filmIsPresent(filmID) & userService.userIsPresent(userID)) {
+            filmService.removeLike(filmID, userID);
+        } else {
+            throw new NoSuchEntryException(Messages.TRY_ANOTHER_ID);
+        }
     }
 
     @GetMapping("/popular")
     public List<Film> getMostLikedFilms(@RequestParam(defaultValue = "10") int count) {
         return filmService.getMostPopularFilms(count);
     }
+
+
 }
